Delete old routes post subcloud network reconfiguration
This commit removes the old routes after subcloud
communication path has been reconfigured e.g.
switch from the management network to the admin network.

Test Plan:
PASS:
1. system host-route-list controller-0
2. check management route to the subcloud
3. dcmanager subcloud update (using admin network parameters)
4. management route removed and admin route created

Story: 2010319
Task: 47315

+++ distributedcloud/dcmanager/common/utils.py
@@ -373,7 +373,7 @@ def delete_subcloud_inventory(inventory_file):
     """Delete the ansible inventory file for the specified subcloud"""
 
     # Delete the file if it exists
-    if os.path.isfile(inventory_file):
+    if inventory_file and os.path.isfile(inventory_file):
         os.remove(inventory_file)
